Staging

+++ internal/modules/warehouse/domain/entities/unit/unit_repository.go
@@ -6,8 +6,8 @@ type Repository interface {
 	Count(ctx context.Context) (int64, error)
 	GetAll(ctx context.Context) ([]*Unit, error)
 	GetPaginated(ctx context.Context, limit, offset int, sortBy []string) ([]*Unit, error)
-	GetByID(ctx context.Context, id int64) (*Unit, error)
+	GetByID(ctx context.Context, id uint) (*Unit, error)
 	Create(ctx context.Context, upload *Unit) error
 	Update(ctx context.Context, upload *Unit) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint) error
 }
